API

Solve

CommonSolve.solve — Method

solve(problem::WaveguideProblem; [Nouts], kwargs...)

Solves the WaveguideProblem, using QuantumOptics.timeevolution.master_nh[_dynamic].

Arguments

problem: the waveguide problem

Nouts::Array{Int} (optional): specifies the dimension of the output Fock spaces

Additional keyword arguments are passed to QuantumOptics.jl's master equation solver, see it's documentation

Returns

ts: The times on which the output is determined.

ρ(t): The density matrix at each point in time. The Hilbert space depends on the specific problem; if unsure check out ρ[1].basis_l.

WaveguideProblem

ChiralWaveguide.WaveguideProblem — Type

WaveguideProblem(system, [ψᵢ::WavePacket, ψ₀::Array{Mode},] ts)

Defines the waveguide problem for solve(x).

Arguments

system: array of form [H, σ, ψₛ] or [H, Ls, σ, ψₛ]. H is the system Hamiltonian, Ls an array of system dissipators, σ the coupling operator of the system, and ψₛ the systems initial state.

ψᵢ (optional): the WavePacket driving the system.

ψₒ (optional): the observed modes (only one possible at the moment).

ts: time domain for the simulation. Can be a number, tuple or array. If ts is a number the simulation is started at t = 0.

WavePackets & States

ChiralWaveguide.WavePacket — Type

WavePacket(mode::Mode, state::WavePacketState)

Describes the input wave packet. A WavePacket consists of any Mode and any WavePacketState.

ChiralWaveguide.WavePacketState — Type

WavePacketState

Abstract base class for all states.

ChiralWaveguide.Coherent — Type

Coherent(α[, N_cutoff])

Describes a coherent state $|α⟩$. The autmatic choice for N_cutoff yields a normalisation of above 0.999 for α ∈ [0,10].

ChiralWaveguide.ArbitraryState — Type

ArbitraryState(amplitudes[, N_cutoff])

Creates a state with specific amplitudes.

ChiralWaveguide.DisplacedArbitraryState — Type

DisplacedArbitraryState(α, amplitudes[, N_cutoff])

Describes a state with specific amplitudes, which then is displaced by $D(α)$.

ChiralWaveguide.Fock — Type

Fock(n[, N_cutoff])

Decribes the Fock state $|n⟩$.

ChiralWaveguide.DisplacedFock — Type

  Fock(n[, N_cutoff])

Decribes the displaced Fock state $D(α)|n⟩$.

ChiralWaveguide.SqueezedVacuum — Type

SqueezedVacuum(r, ϕ[, N_cutoff])

Decribes the a squeezed state with squeezing amplitude r and squeezing angle ϕ, i.e., the state $\exp[(ξ^{*} a^2 - ξ a^{\dagger 2})/2] |0⟩$ with $ξ = r e^{i ϕ}$.

Modes

ChiralWaveguide.Mode — Type

Mode(modeFunction; compression = :algebraic, kwargs...)
Mode(modeFunction, gᵢ, gₒ, norm)

Container for a cavity mode.

Arguments

modeFunction: the wave function of the mode

gᵢ: coupling function for an input cavity

gₒ: coupling function for an output cavity

norm: square integral of the mode up to time t

If only the mode function is provided, then all other quantities are numerically determined. This is done by mapping $ℝ → (-1, 1)$ and then numerically solving the resulting integrals. The argument compression picks the map $ℝ → (-1, 1)$ and can be a Compression or the respective symbol. Allowed symbols are the keys of the ChiralWaveguide.Compressions dictionary. Additional keyword arguments will be passed to the numerical integrator.

ChiralWaveguide.HardBoxMode — Function

HardBoxMode(; t₀ = 0.0, σ = 1.0)

The mode is a box function, i.e., the mode takes the value $1 / \sqrt{σ}$ for $t₀ < t < t₀ + σ$ and 0 otherwise.

ChiralWaveguide.SoftBoxMode — Function

SoftBoxMode(; τ = 0.0, σ = 1.0, n::Int = 10)

The mode is an approximation to the box mode, $\mathrm{box}(t) ≈ \frac{1}{\sqrt{(2t)^{2n} + 1}}$. The parameters τ and σ control the center and width respectively, n the exponent.

ChiralWaveguide.SoftBoxExpMode — Function

SoftBoxExpMode(; τ = 0.0, σ = 1.0, γ = 10.0)

A box mode with exponential decaying flanks, given by

                   ⌜ 1                     for |t| < 1/2
softBoxExp(t, γ) = |
                   ⌞ exp(-γ(|t| - 1/2))    for |t| > 1/2

The parameters τ and σ control the center and width respectively, γ the deay rate.

ChiralWaveguide.GaussMode — Function

GaussMode(; τ = 0.0, σ = 1.0)

Numerical stable mode function & couplings of Gaussian wave packet with mean τ and variance σ.

ChiralWaveguide.ExpMode — Function

ExpMode(; t₀ = 0.0, γ = 1.0)

Exponentially raises up to time t₀ or decays after t₀, depending on the sign of the rate γ. Amplitude decays with rate γ/2, so that the probability density decays with γ.

Systems

ChiralWaveguide.TwoLevelChain — Function

TwoLevelChain(N_atoms::Int; Γ = 0.0)

Creates a chain of N_atoms two-level atoms with spontaneous decay rate Γ. The coupling rate is set to 1, but can be modified by scaling the Hamiltonian and coupling operator. Uses NLevelBasis(2) as the single atom basis.

ChiralWaveguide.DissipativeLambdaChain — Function

DissipativeLambdaChain(N_atoms::Int; γd = 1.0, Γ = 0.0)

Creates a chain of N_atoms three-level atoms with spontaneous decay rate Γ and decay rate γd into the second ground state. The coupling rate is set to 1, but can be modified by scaling the Hamiltonian and coupling operator. Uses NLevelBasis(3) as the single atom basis.

Compressions

ChiralWaveguide.Compression — Type

Compression(c, d, i; σ = 1.0, μ = 0.0)

Maps $ℝ → (-1, 1)$.

Arguments

c: the "compression", i.e., the map $ℝ → (-1, 1)$,

d: the "decompression", i.e., the map $(-1, 1) → ℝ$,

i: the jacobian of the map, i.e., $∂d(τ) /∂τ$,

σ, μ: control the width and offset in the numerical integrations

ChiralWaveguide.Algebraic — Function

Algebraic(; σ = 1.0, μ = 0.0)

Maps $ℝ → (-1, 1)$ by $t ↦ \frac{t}{1 + |t|}$. Equivalent symbol is :algebraic.

ChiralWaveguide.Exponential — Function

Exponential(; σ = 1.0, μ = 0.0)

Maps $ℝ → (-1, 1)$ by $t ↦ \mathrm{tanh}(t)$. Equivalent symbol is :exponential.

ChiralWaveguide.Trigonometric — Function

Trigonometric(; σ = 1.0, μ = 0.0)

Maps $ℝ → (-1, 1)$ by $t ↦ \frac{2}{π} \mathrm{tan}^{-1}(t)$. Equivalent symbol is :trigonometric.